refactor: Centralize agent account types and update references
